Include Swift 5.5’s `_InternalSwiftSyntaxParser.dylib` in the repo
Previously, clients of SwiftSyntax always needed to provide a `_InternalSwiftSyntaxParser.dylib` that was matching the SwiftSyntax version they are using. This was fragile and annoying.

Add the `_InternalSwiftSyntaxParser.dylib` that was released as part of Swift 5.5 to the repo as a binary framework dependency. This way, users can just depend on the SwiftSyntax package and it will bring along the `_InternalSwiftSyntaxParser.dylib` it needs.